monitor: merge EndpointCreateNotification and EndpointDeleteNotification
The types EndpointCreateNotification and EndpointDeleteNotification
contain the same fields. Thus merge them in a single type named
EndpointNotification which is used by func EndpointCreateMessage and
EndpointDeleteMessage.

Because the type is embedded into AgentNotifyMessage the consumer can
still determine whether it was a create or delete event based on
AgentNotifyMessage.Type.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>